modified olden src due to conflicts in malloc original: malloc(unsigned) modified: malloc(size_t)

+++ MultiSource/Benchmarks/Olden/bisort/node.h
@@ -14,6 +14,6 @@ typedef struct future_cell_int{
   HANDLE *value;
 } future_cell_int;
 
-extern void *malloc(unsigned);
+extern void *malloc(size_t);
 
 #define NIL ((HANDLE *) 0)
